Change to GIT complete
The current version now compiles. There's a new DHT module that address the DHT22 directly, this allows for that change. This has NOT been loaded to the coop yet, just tested with verifying the code will compile.

Also uploaded all the older versions so they are in GIT as well.